pyproject.toml
[project]
name = "teacher"
version = "0.1.0"
description = "Teacher pet project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "flake8-import-rules>=1.0.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio==0.21.2",
    "ruff>=0.12.7",
]

[tool.ruff]
target-version = 'py311'
line-length = 120
extend-exclude = [
    "migrations",
]

[tool.ruff.lint]
extend-select = [
    "F",  # PyFlakes
    "E",  # pycodestyle [Error]
    "W", # pycodestyle [Warning]
    "I", # isort
    "YTT", # flake8-2020
    "ASYNC",  # flake8-async
    "F", # flake8-bugbear
    "COM", # flake8-commas
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLOT", # flake8-slots
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "INT", # flake8-gettext
    "PD", # pandas-vet
    "FLY", # flynt
    "LOG", # flake8-logging
]
exclude = [
    "venv",
    ".venv",
    "migrations"
]

isort.lines-after-imports = 2


[tool.ruff.format]
exclude = [
    "venv",
    ".venv",
    "migrations"
]

[tool.mypy]
strict = true
python_version = "3.13"
files = ["src", ]
warn_unused_configs = true
show_error_codes = true
incremental = true
cache_dir = ".mypy_cache"
warn_no_return = true
pretty = true
allow_redefinition = false
exclude = [
    "^\\.git/",
    "^\\.venv/",
    "^__pycache__/",
    "^tests/.*/fixtures/",
]
plugins = ["pydantic.mypy", ]
